pyproject.toml
[project]
name = "despliegue-blockchain-besu"
version = "0.1.0"
description = ""
authors = [
    { name = "Ion Jaureguialzo Sarasola" }
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pyjwt (>=2.12.1,<3.0.0)",
    "cryptography (>=46.0.5,<47.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
